refactor: 이전 코드리뷰 comment 반영
