+++ frontend/techpick/src/features/nodeManagement/api/useCreateFolder.ts
@@ -1,17 +1,9 @@
 import { postFolder } from '@/features/nodeManagement/api/queryFunctions';
 import { useMutation } from '@tanstack/react-query';
+import { ApiFolderData } from '@/shared/types/ApiTypes';
 
 export const useCreateFolder = () => {
-  return useMutation<
-    {
-      id: number;
-      name: string;
-      parentFolderId: number;
-      userId: number;
-    },
-    Error,
-    string
-  >({
+  return useMutation<ApiFolderData, Error, string>({
     mutationFn: postFolder,
   });
 };
