+++ frontend/techpick/src/features/nodeManagement/api/useGetDefaultFolderData.ts
@@ -1,9 +1,9 @@
-import { DefaultFolderIdData } from '@/shared/types/ApiTypes';
+import { ApiDefaultFolderIdData } from '@/shared/types/ApiTypes';
 import { getFoldersIdData } from '@/features/nodeManagement/api/queryFunctions';
 import { useQuery } from '@tanstack/react-query';
 
 export const useGetDefaultFolderData = () => {
-  return useQuery<DefaultFolderIdData>({
+  return useQuery<ApiDefaultFolderIdData>({
     queryKey: ['getDefaultFolderId'],
     queryFn: getFoldersIdData,
   });
