+++ frontend/techpick/src/features/nodeManagement/utils/getNewIdFromStructure.ts
@@ -1,5 +1,5 @@
 import { NodeData } from '@/shared/types';
-import { StructureData } from '@/shared/types/ApiTypes';
+import { ApiStructureData } from '@/shared/types/ApiTypes';
 
 function getMaxIdFromNodes(nodes: NodeData[]): number {
   let maxId = 0;
@@ -19,7 +19,7 @@ function getMaxIdFromNodes(nodes: NodeData[]): number {
   return maxId;
 }
 
-export function getNewIdFromStructure(structure: StructureData): string {
+export function getNewIdFromStructure(structure: ApiStructureData): string {
   const maxIdInRoot = getMaxIdFromNodes(structure.root);
   const maxIdInRecycleBin = getMaxIdFromNodes(structure.recycleBin);
   const maxId = Math.max(maxIdInRoot, maxIdInRecycleBin);
