+++ frontend/techpick/src/features/userManagement/utils/getClientCookie.ts
@@ -1,0 +1,5 @@
+export function getClientCookie(name: string): string | undefined {
+  const cookies = document.cookie.split(';');
+  const cookie = cookies.find((row) => row.startsWith(`${name}`));
+  return cookie ? decodeURIComponent(cookie.split('=')[1]) : undefined;
+}
