+++ frontend/techpick/src/shared/api/apiClient.ts
@@ -1,0 +1,7 @@
+import ky from 'ky';
+
+export const apiClient = ky.create({
+  credentials: 'include',
+  prefixUrl: process.env.NEXT_PUBLIC_API,
+  hooks: {},
+});
