+++ frontend/techpick/src/shared/types/ApiTypes.ts
@@ -1,13 +1,20 @@
 import { NodeData } from '@/shared/types/NodeData';
 
-export interface DefaultFolderIdData {
+export interface ApiDefaultFolderIdData {
   userId: number;
   RECYCLE_BIN: number;
   UNCLASSIFIED: number;
   ROOT: number;
 }
 
-export interface StructureData {
+export interface ApiStructureData {
   root: NodeData[];
   recycleBin: NodeData[];
+}
+
+export interface ApiFolderData {
+  id: number;
+  name: string;
+  parentFolderId: number;
+  userId: number;
 }
